rmm: Enter to realm with rmi::REC_ENTER
rsi::HOST_CALL is used for realm_exit

+++ rmm/armv9a/src/rsi/mod.rs
@@ -1,1 +1,1 @@
-pub const RSI_REMAP_PAGE: u64 = 0xfeed_0001;
+pub const RSI_REMAP_PAGE: usize = 0xfeed_0001;
